Hackathon/requirements.txt
Django == 3.2.12
djangorestframework == 3.13.1
django-cors-headers == 3.13.0
psycopg2-binary 
psycopg2  # for postgres 
django-filter == 22.1
redis # for chaching support with redis 
# tensorflow | for machine learning
#  openai 
#removed apps 
#cors
#auth
#
